Remarkable-employees-db-diagram
I have left two tables with not connection to the others tables because I will be using them as the site grows. This is because the company I work for has properties accross the country therefore the two tables will be relevant as it grows. I will also be adding more tables as the need arises.